feat(view): ページ送りのボタンをクリックしたときに遷移しないようにする
